+++ black_lion/versions.tf
@@ -1,0 +1,11 @@
+terraform {
+  required_providers {
+    docker = {
+      source = "terraform-providers/docker"
+    }
+    template = {
+      source = "hashicorp/template"
+    }
+  }
+  required_version = ">= 0.13"
+}