+++ red_lion/versions.tf
@@ -1,0 +1,14 @@
+terraform {
+  required_providers {
+    docker = {
+      source = "terraform-providers/docker"
+    }
+    random = {
+      source = "hashicorp/random"
+    }
+    template = {
+      source = "hashicorp/template"
+    }
+  }
+  required_version = ">= 0.13"
+}